Revert "Add cairo1-run pretty printing (#1630)"
This reverts commit 061ba872d94a0e49027132f2215b4d4ad4ee95d6.
