+++ CHANGELOG.md
@@ -2,8 +2,6 @@
 
 #### Upcoming Changes
 
-* feat: Add cairo1-run output pretty-printing for felts, arrays/spans and dicts [#1630](https://github.com/lambdaclass/cairo-vm/pull/1630)
-
 * feat: output builtin features for bootloader support [#1580](https://github.com/lambdaclass/cairo-vm/pull/1580)
 
 #### [1.0.0-rc1] - 2024-02-23
